requirements.txt
discord.py==2.5